Fixed bug where query params could not be bound to values.
The QueryString in RequestURI was missing.
Fixed RequestURI recreation in URI2PathAndDontSkip(). Now it works in both
CGI and `serve` modes with QueryString().
Added pager for listing celini in stranici and stranici in root page.
Improved CGI detection and emulation.
Default limit for items in a list is now 10.
Improved default Cfg structure generation.

+++ slovo/helpers.go
@@ -38,12 +38,12 @@ func iHostName(c echo.Context) (host string) {
 }
 
 // Allow only valid values 0,1,2
-func publishedStatus(c echo.Context) uint8 {
+func publishedStatus(c echo.Context) int {
 	preview := c.QueryParam("preview")
 	if preview != "" {
-		return uint8(1)
+		return 1
 	}
-	return uint8(2)
+	return 2
 }
 
 var binDir string
